1.4.10 - 11/20A/2025
- Updated and edited many sections of code and documentation to improve readability and reduce warnings

+++ ftclib/src/main/java/com/epra/epralib/ftclib/math/statistics/Statistics.java
@@ -56,7 +56,7 @@ public class Statistics {
      * @return An array of doubles.*/
     public static double[] toSet(ArrayList<Double> set) {
         double[] r = new double[set.size()];
-        for (int i = 0; i < r.length; i++) { r[i] = set.get(i).doubleValue(); }
+        for (int i = 0; i < r.length; i++) { r[i] = set.get(i); }
         return r;
     }
     /**Converts an ArrayList of longs to an Array.
@@ -64,7 +64,7 @@ public class Statistics {
      * @return An array of longs.*/
     public static long[] toTime(ArrayList<Long> set) {
         long[] r = new long[set.size()];
-        for (int i = 0; i < r.length; i++) { r[i] = set.get(i).longValue(); }
+        for (int i = 0; i < r.length; i++) { r[i] = set.get(i); }
         return r;
     }
 }
